[CP-3185] Minor refactor of device-flash
