+++ libs/msc-flash/msc-flash-harmony/src/lib/selectors/index.ts
@@ -3,7 +3,7 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export * from "./select-flashing-state.selector"
+export * from "./select-flashing-state"
 export * from "./select-flashing-abort-controller"
 export * from "./select-flashing-process-state"
 export * from "./select-is-flashing-in-active-phases"
