+++ libs/shared/utils/src/index.ts
@@ -13,6 +13,7 @@ export * from "./lib/prevent-default-shortcuts"
 export * from "./lib/register-shortcuts"
 export * from "./lib/routes-history.context"
 export * from "./lib/settings-service.interface"
+export * from "./lib/split-path-to-dir-name-and-base-name"
 export * from "./lib/use-debounced-events-handler"
 export * from "./lib/exec-command"
 export * from "./lib/exec-command-with-sudo"
