+++ libs/shared/utils/src/lib/split-path-to-dir-name-and-base-name.ts
@@ -1,0 +1,14 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+import path from "path"
+
+export const splitPathToDirNameAndBaseName = (
+  currentPath: string
+): [string, string] => {
+  const dirName = path.dirname(currentPath)
+  const baseName = path.basename(currentPath)
+  return [dirName, baseName]
+}
